requirements.txt
pandas
numpy
spacy
pdfminer.six
python-docx
scikit-learn
gensim
streamlit
pyngrok
matplotlib
seaborn
